+++ vlib/time/time_nix.c.v
@@ -23,8 +23,8 @@ fn C.timegm(&C.tm) C.time_t
 // fn C.gmtime_r(&tm, &gbuf)
 fn C.localtime_r(t &time_t, tm &C.tm)
 
-fn make_unix_time(t C.tm) int {
-	return int(C.timegm(&t))
+fn make_unix_time(t C.tm) i64 {
+	return i64(C.timegm(&t))
 }
 
 // local returns t with the location set to local time.
